fix import paths after static methods has moved
